src/requirements.txt
azure-cosmos
openai
python-dotenv
tiktoken
flask